Update main.c
Current implementation assumes theta in Equation 4 doesn't decrease with time, therefore there's some overshoot for the inferred angular rate at the start of the measurement where angular rate is larger. However, the results stills calculate the inferred angular rate with high accuracy, since the initial angle is small, therefore the assumption for constant amplitude for theta holds. We can use a second-order damped model to model theta as a function of time to further increase the accuracy for larger angle oscillations.