Add bubble sort test for riscv
This patch adds a sample bubble sort program that
can be compiled with the riscv tool chain and tested
with sunflower-riscv.

It is currently not possiible to print output, so it
is necessary to use debuggers like gdb or the internal
dumpmem instruction, if you know which memory region to
watch over.