Fix uncertain loads and stores
This commit is full of hacks due to limitations in the patching of
the assembler.

`unflw` and `unfsw` have the best guess memory location zeroed out.
Therefore, we use the uncertain memory location to store the best
guess (in normal memory) and the variance (in the uncertain memory
covariance matrix). This is not a long term solution as it requires
a **huge** uncertain memory covariance matrix which it uses very
wastefully. However, the benchmark now runs and is a nice proof of
concept.