Correct parsing of unfsw and unflw instructions.
In d95d0decf84202e2f1b9aae97155ccaffebe720c I worked around the
limitations of the ld patch in by making sunflower parse the
binary that ld generated rather than the correct binary.

With the implementation of a python script that "fixes" the binary,
this commit has sunflower parse the correct binary. This is a much
better approach.